Mostly fleshed out the crm controllers. Finished up positions and tasks

+++ models/crm_position_task.hpp
@@ -3,11 +3,11 @@
 #include "model.hpp"
 #include "model_factory.hpp"
 
-#define CRM_POSITION_TASK_COLUMNS          \
-  COLUMN(position_id, int,     "int")      \
-  COLUMN(task_id,     int,     "int")      \
-  COLUMN(active,      int,     "int")      \
-  COLUMN(created_at,  std::tm, "datetime") \
+#define CRM_POSITION_TASK_COLUMNS           \
+  COLUMN(position_id, long long int, "int") \
+  COLUMN(task_id,     long long int, "int") \
+  COLUMN(active,      int,           "int") \
+  COLUMN(created_at,  std::tm, "datetime")  \
   COLUMN(updated_at,  std::tm, "datetime")
 
 class CrmPositionTask : public Model::Instance<CrmPositionTask> {
